requirements.txt
# Core
python-dotenv>=1.0.0

# Supabase client
supabase>=2.3.0
postgrest>=0.16.0

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# LLM (Gemini)
google-generativeai>=0.5.0

# Utilities
numpy>=1.23.0
requests>=2.31.0


google-genai
spacy
sentence-transformers
torch
